pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zopdev-signals"
version = "0.1.0"
description = "Local-first buying signal tracker for zop.dev"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "typer>=0.12.0",
  "pydantic>=2.7.0",
  "requests>=2.32.0",
  "beautifulsoup4>=4.12.0",
  "feedparser>=6.0.0",
  "rapidfuzz>=3.9.0",
  "python-dateutil>=2.9.0",
  "tenacity>=9.0.0",
  "pandas>=2.2.0",
  "gspread>=6.1.0",
  "google-auth>=2.34.0",
  "pytest>=8.3.0",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
